Added forEach methods
Sometimes it's necessary to have a forEach method (for example,
if you have a large KeyedStream which you can't keep in memory
and need to collect into two data structures.

Here, a forEach method would be useful.

We depart from *Keys for the methods on the keys, because forEach
is clearly a prefix that would be followed by a singular.