Make it so that we can get the futures in source order too
The last PR made it possible to transform a list of futures into
a list of futures in order of completion.

However, it's frequently useful to get the list of futures in source
order with the same parallelism properties too.

This PR adds 'inSourceOrder', which provides the same properties
to the user, but without buffering so aggressively.

It also adds some more stringent tests for MoreStreams, given that
it would be easy to accidentally mistake the two.